Reorganized files
Reorganized files and general clean-up of capabilities.
