+++ bundles/org.openhab.binding.mideaac/src/main/java/org/openhab/binding/mideaac/internal/handler/CommandSet.java
@@ -290,7 +290,7 @@ public class CommandSet extends CommandBase {
         data[0x01] = (byte) 0x0E;
         data[0x09] = (byte) 0x03;
         data[0x0a] = (byte) 0xB5;
-        data[0x0b] = (byte) 0x02;
+        data[0x0b] = (byte) 0x01;
         data[0x0c] = (byte) 0x00;
     }
